Make add_key private
Deprecated in favour of link_dimensions. No longer exposes internal dimension keys, and drops a need to use the AxisDefinition struct (also looking to make private). Will be some performance cost, but that is not judged to be important here.  Method also feels safer, as it checks whether a dimension has been linked before linking it