Update benchmarks following OS update
Benchmarks took quite the hit after going from linux kernal v4.10 to v5.3 due to the meltdown/specter workarounds due to a bug in intel's chips... Benchmarks can be improved to slighly faster than the originals using the boot option pti=off, however it exposes the machine to some of the exploits.  Would be nice to test this on AMD.

Benchmarks were updated without (repo) code changing, as they are there as a relative comparison, and it will skew things if some are updated and others aren't.