Update benchmark results in readme post-refactor
TinyText and NullableF64 4d (populated) stream-read slightly slower due to a bug fix (in the benchmarks) where test values were not  being written to the correct location. Non-TinyText stream-read bencharks all slightly slower as values are now read into a vector instead of an array for consitency's sake (init an array is harder for tiny text as it doesn't implement Copy), the (result) collection instantiation was also moved into the benchmark to allow for simpler benchmark code (at the loss of some accuracy depending on how you see it).